requirements.txt
tensorflow
numpy>=1.24
opencv-python
matplotlib
scikit-learn
pillow
streamlit
facenet-pytorch
tqdm